moved files for About window to mainwindow folder
