+++ src/mainwindow/aboutwindow.cpp
